fix(ios): débloque le build après merge — corrige la PR #265 (éditeur vidéo)
La PR #265 (redesign éditeur vidéo) a été mergée dans un état non
compilable. 3 erreurs réelles corrigées :
- VideoEditSessionStore : propriété actor-isolée référencée depuis une
  autoclosure nonisolated (opérateur ??)
- VideoRenderGeometry : conversion Double -> CGFloat manquante
- VideoEditorViewModel : await manquant sur AVPlayer.seek async

Inclut l'enregistrement pbxproj de ConversationComposerStateTests.

+++ packages/MeeshySDK/Sources/MeeshySDK/Video/VideoRenderGeometry.swift
@@ -62,8 +62,8 @@ public struct VideoRenderGeometry: Sendable, Equatable {
 
         let minX = corners.map(\.x).min() ?? 0
         let minY = corners.map(\.y).min() ?? 0
-        let maxX = corners.map(\.x).max() ?? cropW
-        let maxY = corners.map(\.y).max() ?? cropH
+        let maxX = corners.map(\.x).max() ?? CGFloat(cropW)
+        let maxY = corners.map(\.y).max() ?? CGFloat(cropH)
 
         let normalize = CGAffineTransform(translationX: -minX, y: -minY)
         let total = cropTransform
